backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.8.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Environment & Config
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Background tasks (optional for now)
# celery>=5.0.0
# redis>=4.0.0

# AI Integration
anthropic>=0.3.0
openai>=1.0.0

# HTTP Client
httpx>=0.24.0
requests>=2.25.0

# Development
pytest>=6.0.0
pytest-asyncio>=0.15.0